+++ gotatun/src/noise/rate_limiter.rs
@@ -18,6 +18,7 @@ use crate::packet::{Packet, WgCookieReply, WgHandshakeBase, WgKind};
 use constant_time_eq::constant_time_eq;
 #[cfg(feature = "mock_instant")]
 use mock_instant::thread_local::Instant;
+use rand::TryRngCore;
 use std::collections::HashMap;
 use std::net::IpAddr;
 use std::sync::atomic::{AtomicU64, Ordering};
@@ -30,7 +31,7 @@ use aead::generic_array::GenericArray;
 use aead::{AeadInPlace, KeyInit};
 use chacha20poly1305::{Key, XChaCha20Poly1305};
 use parking_lot::Mutex;
-use rand_core::{OsRng, RngCore};
+use rand::rngs::OsRng;
 
 const COOKIE_REFRESH: u64 = 128; // Use 128 and not 120 so the compiler can optimize out the division
 const COOKIE_SIZE: usize = 16;
@@ -79,7 +80,7 @@ impl RateLimiter {
     /// * `limit` - Maximum number of packets allowed per rate limiting period
     pub fn new(public_key: &crate::x25519::PublicKey, limit: u64) -> Self {
         let mut secret_key = [0u8; 16];
-        OsRng.fill_bytes(&mut secret_key);
+        OsRng.try_fill_bytes(&mut secret_key).unwrap();
         RateLimiter {
             nonce_key: Self::rand_bytes(),
             secret_key,
@@ -97,7 +98,7 @@ impl RateLimiter {
 
     fn rand_bytes() -> [u8; 32] {
         let mut key = [0u8; 32];
-        OsRng.fill_bytes(&mut key);
+        OsRng.try_fill_bytes(&mut key).unwrap();
         key
     }